chore(naming): refactor to adhere to style guide
