+++ client/components/FlashMessagesList/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import FlashMessage from './FlashMessage';
+import FlashMessage from './FlashMessage.jsx';
 import { deleteFlashMessage } from '../../actions/flashMessages';
 
 class FlashMessagesList extends React.Component {
