+++ client/components/LoginPage/LoginForm.jsx
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
 import toastr from 'toastr';
-import TextFieldGroup from '../common/TextFieldGroup';
+import TextFieldGroup from '../Common/TextFieldGroup.jsx';
 import validateInput from '../../../server/shared/validations/login';
 import { login } from '../../actions/authenticationAction';
 import { addFlashMessage } from '../../actions/flashMessages';
