+++ client/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { reducer } from 'react-redux-sweetalert';
-import flashMessages from './flashMessages';
+import flashMessages from './flashMessagesReducer';
 import handleUsers from './userReducer';
 import handleDocuments from './documentReducer';
 import manageRoles from './roleReducer';
