+++ client/routes.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 import App from './components/App';
-import HomePage from './components/home/HomePage';
-import AboutPage from './components/about/AboutPage';
-import SignupPage from './components/authentication/SignupPage';
-import LoginPage from './components/login/LoginPage';
-import DocumentPage from './components/document/DocumentPage';
-import DashboardPage from './components/dashboard/DashboardPage';
-import ManangeRolePage from './components/adminComponents/ManageRolePage';
-import HandleUsersPage from './components/adminComponents/HandleUsersPage';
-import ProfilePage from './components/ProfilePage/ProfilePage';
-import requireAuth from './utils/requireAuth';
-import requireAdminAuth from './utils/requireAdminAuth';
+import HomePage from './components/HomePage/index.jsx';
+import AboutPage from './components/AboutPage/index.jsx';
+import SignupPage from './components/SignupPage/index.jsx';
+import LoginPage from './components/LoginPage/index.jsx';
+import DocumentPage from './components/DocumentPage/index.jsx';
+import DashboardPage from './components/DashboardPage/index.jsx';
+import ManangeRolePage from './components/adminComponents/ManageRolePage.jsx';
+import HandleUsersPage from './components/adminComponents/HandleUsersPage.jsx';
+import ProfilePage from './components/ProfilePage/index.jsx';
+import requireAuth from './Utils/RequireAuth.jsx';
+import requireAdminAuth from './Utils/RequireAdminAuth.jsx';
 
 export default (
   <Route path="/" component={App}>
